タイトルシーン修正

+++ Project/ApplicationLayer/Enemy/Enemy.cpp
@@ -146,7 +146,7 @@ void Enemy::OnCollision(Collider* other)
 	}
 }
 
-void Enemy::TakeDamage(int damage)
+void Enemy::TakeDamage(float damage)
 {
 	if (isDead_) return;
 	hp_ -= damage;
